\name{indonRespir}
\docType{data}
\alias{indonRespir}
\title{The Indonesian Children's Health Study data}
\description{The Indonesian Children's Health Study data, corresponding to a cohort of 275 Indonesian children. The response variable is respiratory infection whilst potential predictors are age, indicator of vitamin A deficiency, sex, height, indicator of being stunted and the number of clinic visits for each child.}
